refactor: 프로필 api 통합
프로필을 가져오는 api와 프로필을 수정하는 api를 통합함.
GET방식으로 query에 userId를 담아 보내면 프로필을 가져오고, POST방식으로 요청을 넣으면 프로필을 수정한다.
